refactor: use `AuthenticationMethod` from CSL
