+++ security/security-core/src/main/java/io/camunda/security/configuration/SecurityConfigurations.java
@@ -7,7 +7,7 @@
  */
 package io.camunda.security.configuration;
 
-import io.camunda.security.entity.AuthenticationMethod;
+import io.camunda.security.api.model.config.AuthenticationMethod;
 
 public class SecurityConfigurations {
   public static SecurityConfiguration unauthenticatedAndUnauthorized() {
