+++ tasklist/webapp/pom.xml
@@ -31,6 +31,11 @@
 
     <dependency>
       <groupId>io.camunda</groupId>
+      <artifactId>camunda-security-library-api</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>io.camunda</groupId>
       <artifactId>camunda-spring-utils</artifactId>
     </dependency>
