+++ zeebe/qa/integration-tests/src/test/java/io/camunda/zeebe/it/engine/authorization/BasicAuthOverGrpcIT.java
@@ -17,7 +17,7 @@ import io.camunda.client.api.command.ClientStatusException;
 import io.camunda.client.api.search.enums.PermissionType;
 import io.camunda.client.api.search.enums.ResourceType;
 import io.camunda.configuration.SecondaryStorage.SecondaryStorageType;
-import io.camunda.security.entity.AuthenticationMethod;
+import io.camunda.security.api.model.config.AuthenticationMethod;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
 import io.camunda.zeebe.model.bpmn.Bpmn;
