+++ zeebe/qa/integration-tests/src/test/java/io/camunda/zeebe/it/engine/authorization/OidcAuthOverRestStartupIT.java
@@ -7,7 +7,7 @@
  */
 package io.camunda.zeebe.it.engine.authorization;
 
-import io.camunda.security.entity.AuthenticationMethod;
+import io.camunda.security.api.model.config.AuthenticationMethod;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
